[PIDM-590] fix: resolved various bugs on RE storage

+++ src/main/java/it/gov/pagopa/fdr/repository/entity/re/BlobRefEntity.java
@@ -11,5 +11,5 @@ public class BlobRefEntity {
 
   private String fileName;
 
-  private Long fileLength;
+  private Integer fileLength;
 }
